requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.21.0
Pillow>=8.0.0
scikit-learn>=0.24.0
timm>=0.6.0
flask>=2.0.0
werkzeug>=2.0.0 